Update and rename batch_download.py to downloader.py
Implement proper error handling of network errors. 

Implementing a sq3lite database to make it easier to stop/resume downloading.
